Extract AnnotationCanvas from AnnotationRenderService (#49 phase 2)
Canvas rendering, DPR resize, visual feedback element, and spinner
move into a dedicated AnnotationCanvas class. The service delegates
all DOM/canvas concerns via this.canvas.* calls.
